+++ src/Housing.Forecast.Context/ForecastContext.cs
@@ -5,8 +5,8 @@ using System.Text;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.ChangeTracking;
 using Microsoft.Extensions.Logging;
-using Housing.Forecast.Library.Models;
-using Housing.Forecast.Library;
+using Housing.Forecast.Context.Models;
+using Housing.Forecast.Context;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.AspNetCore.Hosting;
