+++ src/Housing.Forecast.Context/Repos/RoomRepo.cs
@@ -2,8 +2,8 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
-using Housing.Forecast.Library;
-using Housing.Forecast.Library.Models;
+using Housing.Forecast.Context;
+using Housing.Forecast.Context.Models;
 
 namespace Housing.Forecast.Context.Repos
 {
