+++ src/Housing.Forecast.Service/Controllers/ForecastController.cs
@@ -7,7 +7,7 @@ using Housing.Forecast.Context;
 using Housing.Forecast.Context.Repos;
 using System.Collections.Generic;
 using System.Linq;
-using Housing.Forecast.Library.Models;
+using Housing.Forecast.Context.Models;
 
 namespace Housing.Forecast.Service.Controllers
 {
