+++ src/Housing.Forecast.Testing/Context/UserRepoTests.cs
@@ -7,10 +7,10 @@ using Microsoft.EntityFrameworkCore;
 
 using Housing.Forecast.Context;
 using Housing.Forecast.Context.Repos;
-using Housing.Forecast.Library.Models;
+using Housing.Forecast.Context.Models;
 using Housing.Forecast.Service;
 
-namespace Housing.Forecast.Testing.Library
+namespace Housing.Forecast.Testing.Context
 {
     public class UserRepoTests
     {
